feat(module): implemented module-like structure with additional support for customization of variants and videos
did a lot of refactoring mostly and working towards improving the way the code interacts with signingsavvy as a platform so it can work like an unofficial API. I did a lot of abstraction with regards to pysign__videos.py and pysign__variants.py, as those were ones that I felt were getting quite repetitive. The only thing that I really need to do at some point is make sure that I write documentation for all of these functions. Although, I am eager to get these flashcards created, I need to take the time to measure twice and cut once.